// src/app/components/core/models/fabric-details.model.ts
export interface FabricDetail {
  refNo?: string;
  fabName?: string;
  fabType?: string;
  quality?: string;
  count?: string;
  construction?: string;
  gsm?: number;
  color?: string;
  uom?: string;
  quantity?: number;
  width?: string;
  length?: number;
  costPrice?: number;
  sellPrice?: number;
  createdDate?: string;
}